controller updated

+++ backend/controllers/product.controller.js
@@ -43,11 +43,14 @@ export const UpdateProduct = async (req, res) => {
 
 export const DeleteProduct = async (req, res) => {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).send('No product with that id');
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted"});
     } catch (error) {
         console.log(`Error in delete product: ${error.message}`);
-        res.status(404).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server error"});
     }
 };
